@@ -2,50 +2,49 @@
   "_args": [
     [
       {
-        "raw": "@angular/animations@4.3.0",
+        "raw": "@angular/animations@^4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fanimations",
         "name": "@angular/animations",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
-        "type": "version"
+        "rawSpec": "^4.4.6",
+        "spec": ">=4.4.6 <5.0.0",
+        "type": "range"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/animations@4.3.0",
-  "_id": "@angular/animations@4.3.0",
+  "_from": "@angular/animations@>=4.4.6 <5.0.0",
+  "_id": "@angular/animations@4.4.6",
   "_inCache": true,
   "_location": "/@angular/animations",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/animations-4.3.0.tgz_1500063673222_0.7006766661070287"
+    "tmp": "tmp/animations-4.4.6.tgz_1508372338929_0.6808346090838313"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/animations@4.3.0",
+    "raw": "@angular/animations@^4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fanimations",
     "name": "@angular/animations",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
-    "type": "version"
+    "rawSpec": "^4.4.6",
+    "spec": ">=4.4.6 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
-    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/animations/-/animations-4.3.0.tgz",
-  "_shasum": "56f34b84649379202ac359929b82eb0b915e9c72",
+  "_resolved": "https://registry.npmjs.org/@angular/animations/-/animations-4.4.6.tgz",
+  "_shasum": "fa661899a8a4e38cb7c583c7a5c97ce65d592a35",
   "_shrinkwrap": null,
-  "_spec": "@angular/animations@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/animations@^4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "author": {
     "name": "angular"
   },
@@ -59,8 +58,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "56f34b84649379202ac359929b82eb0b915e9c72",
-    "tarball": "https://registry.npmjs.org/@angular/animations/-/animations-4.3.0.tgz"
+    "shasum": "fa661899a8a4e38cb7c583c7a5c97ce65d592a35",
+    "tarball": "https://registry.npmjs.org/@angular/animations/-/animations-4.4.6.tgz"
   },
   "es2015": "./@angular/animations.js",
   "homepage": "https://github.com/angular/angular#readme",
@@ -76,7 +75,7 @@
   "name": "@angular/animations",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/core": "4.3.0"
+    "@angular/core": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -85,5 +84,5 @@
   },
   "scripts": {},
   "typings": "./animations.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
