@@ -2,50 +2,50 @@
   "_args": [
     [
       {
-        "raw": "@angular/cdk",
+        "raw": "@angular/cdk@2.0.0-beta.12",
         "scope": "@angular",
         "escapedName": "@angular%2fcdk",
         "name": "@angular/cdk",
-        "rawSpec": "",
-        "spec": "latest",
-        "type": "tag"
+        "rawSpec": "2.0.0-beta.12",
+        "spec": "2.0.0-beta.12",
+        "type": "version"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/cdk@latest",
-  "_id": "@angular/cdk@2.0.0-beta.8",
+  "_from": "@angular/cdk@2.0.0-beta.12",
+  "_id": "@angular/cdk@2.0.0-beta.12",
   "_inCache": true,
   "_location": "/@angular/cdk",
-  "_nodeVersion": "6.10.3",
+  "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/cdk-2.0.0-beta.8.tgz_1499360906424_0.5165699042845517"
+    "tmp": "tmp/cdk-2.0.0-beta.12.tgz_1507246798151_0.8973107798956335"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "5.0.4",
+  "_npmVersion": "3.10.10",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/cdk",
+    "raw": "@angular/cdk@2.0.0-beta.12",
     "scope": "@angular",
     "escapedName": "@angular%2fcdk",
     "name": "@angular/cdk",
-    "rawSpec": "",
-    "spec": "latest",
-    "type": "tag"
+    "rawSpec": "2.0.0-beta.12",
+    "spec": "2.0.0-beta.12",
+    "type": "version"
   },
   "_requiredBy": [
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/cdk/-/cdk-2.0.0-beta.8.tgz",
-  "_shasum": "71961c851dfbeb19e085e898bf5e4461408f8b57",
+  "_resolved": "https://registry.npmjs.org/@angular/cdk/-/cdk-2.0.0-beta.12.tgz",
+  "_shasum": "3a243cb62b93f4e039120ba70f900dc9e235622e",
   "_shrinkwrap": null,
-  "_spec": "@angular/cdk",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/cdk@2.0.0-beta.12",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "bugs": {
     "url": "https://github.com/angular/material2/issues"
   },
@@ -56,11 +56,10 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "integrity": "sha512-OOtK+AA14cmRG9AbUgvoKC9Tooz0N37GTaRSV+xziC8GxXHgwvTu4PFSFHlBnHPipOYC/tB2oP39j3KuurEMPA==",
-    "shasum": "71961c851dfbeb19e085e898bf5e4461408f8b57",
-    "tarball": "https://registry.npmjs.org/@angular/cdk/-/cdk-2.0.0-beta.8.tgz"
+    "shasum": "3a243cb62b93f4e039120ba70f900dc9e235622e",
+    "tarball": "https://registry.npmjs.org/@angular/cdk/-/cdk-2.0.0-beta.12.tgz"
   },
-  "es2015": "./@angular/cdkjs",
+  "es2015": "./esm2015/cdk.js",
   "homepage": "https://github.com/angular/material2#readme",
   "keywords": [
     "angular",
@@ -77,18 +76,19 @@
       "email": "angular-core+npm@google.com"
     }
   ],
-  "module": "./@angular/cdk.es5.js",
+  "module": "./esm5/cdk.es5.js",
   "name": "@angular/cdk",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/core": "^4.0.0",
-    "@angular/common": "^4.0.0"
+    "@angular/core": "~4.4.4",
+    "@angular/common": "~4.4.4"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/angular/material2.git"
   },
+  "scripts": {},
   "typings": "./cdk.d.ts",
-  "version": "2.0.0-beta.8"
+  "version": "2.0.0-beta.12"
 }
