@@ -2,50 +2,50 @@
   "_args": [
     [
       {
-        "raw": "@angular/common@4.3.0",
+        "raw": "@angular/common@4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fcommon",
         "name": "@angular/common",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
+        "rawSpec": "4.4.6",
+        "spec": "4.4.6",
         "type": "version"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/common@4.3.0",
-  "_id": "@angular/common@4.3.0",
+  "_from": "@angular/common@4.4.6",
+  "_id": "@angular/common@4.4.6",
   "_inCache": true,
   "_location": "/@angular/common",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/common-4.3.0.tgz_1500063675025_0.9408791256137192"
+    "tmp": "tmp/common-4.4.6.tgz_1508372341681_0.8242098202463239"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/common@4.3.0",
+    "raw": "@angular/common@4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fcommon",
     "name": "@angular/common",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
+    "rawSpec": "4.4.6",
+    "spec": "4.4.6",
     "type": "version"
   },
   "_requiredBy": [
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/common/-/common-4.3.0.tgz",
-  "_shasum": "13a54a6929dd52f9729b16ae446fad58fe163053",
+  "_resolved": "https://registry.npmjs.org/@angular/common/-/common-4.4.6.tgz",
+  "_shasum": "4b81420724e0828a0e839b95a55eb1a7e83918f2",
   "_shrinkwrap": null,
-  "_spec": "@angular/common@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/common@4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "author": {
     "name": "angular"
   },
@@ -59,8 +59,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "13a54a6929dd52f9729b16ae446fad58fe163053",
-    "tarball": "https://registry.npmjs.org/@angular/common/-/common-4.3.0.tgz"
+    "shasum": "4b81420724e0828a0e839b95a55eb1a7e83918f2",
+    "tarball": "https://registry.npmjs.org/@angular/common/-/common-4.4.6.tgz"
   },
   "es2015": "./@angular/common.js",
   "homepage": "https://github.com/angular/angular#readme",
@@ -76,7 +76,7 @@
   "name": "@angular/common",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/core": "4.3.0"
+    "@angular/core": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -85,5 +85,5 @@
   },
   "scripts": {},
   "typings": "./common.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
