@@ -2,50 +2,49 @@
   "_args": [
     [
       {
-        "raw": "@angular/compiler-cli@4.3.0",
+        "raw": "@angular/compiler-cli@^4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fcompiler-cli",
         "name": "@angular/compiler-cli",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
-        "type": "version"
+        "rawSpec": "^4.4.6",
+        "spec": ">=4.4.6 <5.0.0",
+        "type": "range"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/compiler-cli@4.3.0",
-  "_id": "@angular/compiler-cli@4.3.0",
+  "_from": "@angular/compiler-cli@>=4.4.6 <5.0.0",
+  "_id": "@angular/compiler-cli@4.4.6",
   "_inCache": true,
   "_location": "/@angular/compiler-cli",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/compiler-cli-4.3.0.tgz_1500063679541_0.18050106521695852"
+    "tmp": "tmp/compiler-cli-4.4.6.tgz_1508372348976_0.5218051576521248"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/compiler-cli@4.3.0",
+    "raw": "@angular/compiler-cli@^4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fcompiler-cli",
     "name": "@angular/compiler-cli",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
-    "type": "version"
+    "rawSpec": "^4.4.6",
+    "spec": ">=4.4.6 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
-    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/compiler-cli/-/compiler-cli-4.3.0.tgz",
-  "_shasum": "f375809730f5e883cfe211ae991210f1cb1a9f1e",
+  "_resolved": "https://registry.npmjs.org/@angular/compiler-cli/-/compiler-cli-4.4.6.tgz",
+  "_shasum": "bafd3d1e260e99087eb9a8cf7532dbd603abb9b1",
   "_shrinkwrap": null,
-  "_spec": "@angular/compiler-cli@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/compiler-cli@^4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "bin": {
     "ngc": "./src/main.js",
     "ng-xi18n": "./src/extract_i18n.js"
@@ -66,7 +65,7 @@
     }
   ],
   "dependencies": {
-    "@angular/tsc-wrapped": "4.3.0",
+    "@angular/tsc-wrapped": "4.4.6",
     "minimist": "^1.2.0",
     "reflect-metadata": "^0.1.2"
   },
@@ -74,8 +73,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "f375809730f5e883cfe211ae991210f1cb1a9f1e",
-    "tarball": "https://registry.npmjs.org/@angular/compiler-cli/-/compiler-cli-4.3.0.tgz"
+    "shasum": "bafd3d1e260e99087eb9a8cf7532dbd603abb9b1",
+    "tarball": "https://registry.npmjs.org/@angular/compiler-cli/-/compiler-cli-4.4.6.tgz"
   },
   "homepage": "https://github.com/angular/angular/tree/master/packages/compiler-cli",
   "keywords": [
@@ -94,8 +93,8 @@
   "optionalDependencies": {},
   "peerDependencies": {
     "typescript": "^2.0.2",
-    "@angular/compiler": "4.3.0",
-    "@angular/core": "4.3.0"
+    "@angular/compiler": "4.4.6",
+    "@angular/core": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -104,5 +103,5 @@
   },
   "scripts": {},
   "typings": "index.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
