@@ -2,50 +2,49 @@
   "_args": [
     [
       {
-        "raw": "@angular/compiler@4.3.0",
+        "raw": "@angular/compiler@^4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fcompiler",
         "name": "@angular/compiler",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
-        "type": "version"
+        "rawSpec": "^4.4.6",
+        "spec": ">=4.4.6 <5.0.0",
+        "type": "range"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/compiler@4.3.0",
-  "_id": "@angular/compiler@4.3.0",
+  "_from": "@angular/compiler@>=4.4.6 <5.0.0",
+  "_id": "@angular/compiler@4.4.6",
   "_inCache": true,
   "_location": "/@angular/compiler",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/compiler-4.3.0.tgz_1500063677836_0.4292892280500382"
+    "tmp": "tmp/compiler-4.4.6.tgz_1508372347211_0.6657238877378404"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/compiler@4.3.0",
+    "raw": "@angular/compiler@^4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fcompiler",
     "name": "@angular/compiler",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
-    "type": "version"
+    "rawSpec": "^4.4.6",
+    "spec": ">=4.4.6 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
-    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/compiler/-/compiler-4.3.0.tgz",
-  "_shasum": "55503bf27a1f062f71b9495393f3311903a8fc43",
+  "_resolved": "https://registry.npmjs.org/@angular/compiler/-/compiler-4.4.6.tgz",
+  "_shasum": "2ee1faf25b757e1d128979074be7fae529b3bc20",
   "_shrinkwrap": null,
-  "_spec": "@angular/compiler@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/compiler@^4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "author": {
     "name": "angular"
   },
@@ -59,8 +58,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "55503bf27a1f062f71b9495393f3311903a8fc43",
-    "tarball": "https://registry.npmjs.org/@angular/compiler/-/compiler-4.3.0.tgz"
+    "shasum": "2ee1faf25b757e1d128979074be7fae529b3bc20",
+    "tarball": "https://registry.npmjs.org/@angular/compiler/-/compiler-4.4.6.tgz"
   },
   "es2015": "./@angular/compiler.js",
   "homepage": "https://github.com/angular/angular#readme",
@@ -76,7 +75,7 @@
   "name": "@angular/compiler",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/core": "4.3.0"
+    "@angular/core": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -85,5 +84,5 @@
   },
   "scripts": {},
   "typings": "./compiler.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
