@@ -2,50 +2,49 @@
   "_args": [
     [
       {
-        "raw": "@angular/forms@4.3.0",
+        "raw": "@angular/forms@^4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fforms",
         "name": "@angular/forms",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
-        "type": "version"
+        "rawSpec": "^4.4.6",
+        "spec": ">=4.4.6 <5.0.0",
+        "type": "range"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/forms@4.3.0",
-  "_id": "@angular/forms@4.3.0",
+  "_from": "@angular/forms@>=4.4.6 <5.0.0",
+  "_id": "@angular/forms@4.4.6",
   "_inCache": true,
   "_location": "/@angular/forms",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/forms-4.3.0.tgz_1500063683789_0.12468657363206148"
+    "tmp": "tmp/forms-4.4.6.tgz_1508372355219_0.36747080576606095"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/forms@4.3.0",
+    "raw": "@angular/forms@^4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fforms",
     "name": "@angular/forms",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
-    "type": "version"
+    "rawSpec": "^4.4.6",
+    "spec": ">=4.4.6 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
-    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/forms/-/forms-4.3.0.tgz",
-  "_shasum": "7d0c7a854737e9a30a5fd9665f8d4f56a1b91bd8",
+  "_resolved": "https://registry.npmjs.org/@angular/forms/-/forms-4.4.6.tgz",
+  "_shasum": "fe64ace42435c1b80f49034b7c41ce8caf14a44a",
   "_shrinkwrap": null,
-  "_spec": "@angular/forms@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/forms@^4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "author": {
     "name": "angular"
   },
@@ -59,8 +58,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "7d0c7a854737e9a30a5fd9665f8d4f56a1b91bd8",
-    "tarball": "https://registry.npmjs.org/@angular/forms/-/forms-4.3.0.tgz"
+    "shasum": "fe64ace42435c1b80f49034b7c41ce8caf14a44a",
+    "tarball": "https://registry.npmjs.org/@angular/forms/-/forms-4.4.6.tgz"
   },
   "es2015": "./@angular/forms.js",
   "homepage": "https://github.com/angular/angular#readme",
@@ -76,9 +75,9 @@
   "name": "@angular/forms",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/core": "4.3.0",
-    "@angular/common": "4.3.0",
-    "@angular/platform-browser": "4.3.0"
+    "@angular/core": "4.4.6",
+    "@angular/common": "4.4.6",
+    "@angular/platform-browser": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -87,5 +86,5 @@
   },
   "scripts": {},
   "typings": "./forms.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
