@@ -2,50 +2,49 @@
   "_args": [
     [
       {
-        "raw": "@angular/http@4.3.0",
+        "raw": "@angular/http@^4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fhttp",
         "name": "@angular/http",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
-        "type": "version"
+        "rawSpec": "^4.4.6",
+        "spec": ">=4.4.6 <5.0.0",
+        "type": "range"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/http@4.3.0",
-  "_id": "@angular/http@4.3.0",
+  "_from": "@angular/http@>=4.4.6 <5.0.0",
+  "_id": "@angular/http@4.4.6",
   "_inCache": true,
   "_location": "/@angular/http",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/http-4.3.0.tgz_1500063685340_0.19398058485239744"
+    "tmp": "tmp/http-4.4.6.tgz_1508372357342_0.20307665155269206"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/http@4.3.0",
+    "raw": "@angular/http@^4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fhttp",
     "name": "@angular/http",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
-    "type": "version"
+    "rawSpec": "^4.4.6",
+    "spec": ">=4.4.6 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
-    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/http/-/http-4.3.0.tgz",
-  "_shasum": "dfb73310a840a6ad8050ac51f0e55c4984db0926",
+  "_resolved": "https://registry.npmjs.org/@angular/http/-/http-4.4.6.tgz",
+  "_shasum": "0af680c6710bdc026d940e225cfd0f6a5c005d0c",
   "_shrinkwrap": null,
-  "_spec": "@angular/http@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/http@^4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "author": {
     "name": "angular"
   },
@@ -59,8 +58,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "dfb73310a840a6ad8050ac51f0e55c4984db0926",
-    "tarball": "https://registry.npmjs.org/@angular/http/-/http-4.3.0.tgz"
+    "shasum": "0af680c6710bdc026d940e225cfd0f6a5c005d0c",
+    "tarball": "https://registry.npmjs.org/@angular/http/-/http-4.4.6.tgz"
   },
   "es2015": "./@angular/http.js",
   "homepage": "https://github.com/angular/angular#readme",
@@ -77,8 +76,8 @@
   "optionalDependencies": {},
   "peerDependencies": {
     "rxjs": "^5.0.1",
-    "@angular/core": "4.3.0",
-    "@angular/platform-browser": "4.3.0"
+    "@angular/core": "4.4.6",
+    "@angular/platform-browser": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -87,5 +86,5 @@
   },
   "scripts": {},
   "typings": "./http.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
