@@ -2,50 +2,50 @@
   "_args": [
     [
       {
-        "raw": "@angular/material@^2.0.0-beta.8",
+        "raw": "@angular/material@2.0.0-beta.12",
         "scope": "@angular",
         "escapedName": "@angular%2fmaterial",
         "name": "@angular/material",
-        "rawSpec": "^2.0.0-beta.8",
-        "spec": ">=2.0.0-beta.8 <3.0.0",
-        "type": "range"
+        "rawSpec": "2.0.0-beta.12",
+        "spec": "2.0.0-beta.12",
+        "type": "version"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/material@>=2.0.0-beta.8 <3.0.0",
-  "_id": "@angular/material@2.0.0-beta.8",
+  "_from": "@angular/material@2.0.0-beta.12",
+  "_id": "@angular/material@2.0.0-beta.12",
   "_inCache": true,
   "_location": "/@angular/material",
-  "_nodeVersion": "6.10.3",
+  "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/material-2.0.0-beta.8.tgz_1499360909187_0.62446209625341"
+    "tmp": "tmp/material-2.0.0-beta.12.tgz_1507246802163_0.14236839744262397"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "5.0.4",
+  "_npmVersion": "3.10.10",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/material@^2.0.0-beta.8",
+    "raw": "@angular/material@2.0.0-beta.12",
     "scope": "@angular",
     "escapedName": "@angular%2fmaterial",
     "name": "@angular/material",
-    "rawSpec": "^2.0.0-beta.8",
-    "spec": ">=2.0.0-beta.8 <3.0.0",
-    "type": "range"
+    "rawSpec": "2.0.0-beta.12",
+    "spec": "2.0.0-beta.12",
+    "type": "version"
   },
   "_requiredBy": [
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/material/-/material-2.0.0-beta.8.tgz",
-  "_shasum": "a92852abc9261aea26c2401f576645470be2cf38",
+  "_resolved": "https://registry.npmjs.org/@angular/material/-/material-2.0.0-beta.12.tgz",
+  "_shasum": "71b6d0b7b021891e5d0e3688c1d4bd78c7457f58",
   "_shrinkwrap": null,
-  "_spec": "@angular/material@^2.0.0-beta.8",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/material@2.0.0-beta.12",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "bugs": {
     "url": "https://github.com/angular/material2/issues"
   },
@@ -56,11 +56,10 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "integrity": "sha512-4+OecvjU15i+l/vXBP2qEHdlsU9taK6kBhsWKsxNLK3+TAVoV5qjc2rjOucHtTwI/oOjyBXnLJP6pl4tuLEUQw==",
-    "shasum": "a92852abc9261aea26c2401f576645470be2cf38",
-    "tarball": "https://registry.npmjs.org/@angular/material/-/material-2.0.0-beta.8.tgz"
+    "shasum": "71b6d0b7b021891e5d0e3688c1d4bd78c7457f58",
+    "tarball": "https://registry.npmjs.org/@angular/material/-/material-2.0.0-beta.12.tgz"
   },
-  "es2015": "./@angular/material.js",
+  "es2015": "./esm2015/material.js",
   "homepage": "https://github.com/angular/material2#readme",
   "keywords": [
     "angular",
@@ -76,19 +75,20 @@
       "email": "angular-core+npm@google.com"
     }
   ],
-  "module": "./@angular/material.es5.js",
+  "module": "./esm5/material.es5.js",
   "name": "@angular/material",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/cdk": "2.0.0-beta.8",
-    "@angular/core": "^4.0.0",
-    "@angular/common": "^4.0.0"
+    "@angular/cdk": "2.0.0-beta.12",
+    "@angular/core": "~4.4.4",
+    "@angular/common": "~4.4.4"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/angular/material2.git"
   },
+  "scripts": {},
   "typings": "./material.d.ts",
-  "version": "2.0.0-beta.8"
+  "version": "2.0.0-beta.12"
 }
