@@ -2,50 +2,49 @@
   "_args": [
     [
       {
-        "raw": "@angular/platform-browser-dynamic@4.3.0",
+        "raw": "@angular/platform-browser-dynamic@^4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fplatform-browser-dynamic",
         "name": "@angular/platform-browser-dynamic",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
-        "type": "version"
+        "rawSpec": "^4.4.6",
+        "spec": ">=4.4.6 <5.0.0",
+        "type": "range"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/platform-browser-dynamic@4.3.0",
-  "_id": "@angular/platform-browser-dynamic@4.3.0",
+  "_from": "@angular/platform-browser-dynamic@>=4.4.6 <5.0.0",
+  "_id": "@angular/platform-browser-dynamic@4.4.6",
   "_inCache": true,
   "_location": "/@angular/platform-browser-dynamic",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/platform-browser-dynamic-4.3.0.tgz_1500063690994_0.08206917135976255"
+    "tmp": "tmp/platform-browser-dynamic-4.4.6.tgz_1508372367454_0.036063374718651175"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/platform-browser-dynamic@4.3.0",
+    "raw": "@angular/platform-browser-dynamic@^4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fplatform-browser-dynamic",
     "name": "@angular/platform-browser-dynamic",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
-    "type": "version"
+    "rawSpec": "^4.4.6",
+    "spec": ">=4.4.6 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
-    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/platform-browser-dynamic/-/platform-browser-dynamic-4.3.0.tgz",
-  "_shasum": "551fb18851b27ee8f3e4b0ee25aad10bd7b312e3",
+  "_resolved": "https://registry.npmjs.org/@angular/platform-browser-dynamic/-/platform-browser-dynamic-4.4.6.tgz",
+  "_shasum": "4d3d9a6a7bf2cf3de4058a615ae059eff641fa36",
   "_shrinkwrap": null,
-  "_spec": "@angular/platform-browser-dynamic@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/platform-browser-dynamic@^4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "author": {
     "name": "angular"
   },
@@ -59,8 +58,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "551fb18851b27ee8f3e4b0ee25aad10bd7b312e3",
-    "tarball": "https://registry.npmjs.org/@angular/platform-browser-dynamic/-/platform-browser-dynamic-4.3.0.tgz"
+    "shasum": "4d3d9a6a7bf2cf3de4058a615ae059eff641fa36",
+    "tarball": "https://registry.npmjs.org/@angular/platform-browser-dynamic/-/platform-browser-dynamic-4.4.6.tgz"
   },
   "es2015": "./@angular/platform-browser-dynamic.js",
   "homepage": "https://github.com/angular/angular#readme",
@@ -76,10 +75,10 @@
   "name": "@angular/platform-browser-dynamic",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/core": "4.3.0",
-    "@angular/common": "4.3.0",
-    "@angular/compiler": "4.3.0",
-    "@angular/platform-browser": "4.3.0"
+    "@angular/core": "4.4.6",
+    "@angular/common": "4.4.6",
+    "@angular/compiler": "4.4.6",
+    "@angular/platform-browser": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -88,5 +87,5 @@
   },
   "scripts": {},
   "typings": "./platform-browser-dynamic.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
