@@ -2,50 +2,49 @@
   "_args": [
     [
       {
-        "raw": "@angular/platform-browser@4.3.0",
+        "raw": "@angular/platform-browser@^4.4.6",
         "scope": "@angular",
         "escapedName": "@angular%2fplatform-browser",
         "name": "@angular/platform-browser",
-        "rawSpec": "4.3.0",
-        "spec": "4.3.0",
-        "type": "version"
+        "rawSpec": "^4.4.6",
+        "spec": ">=4.4.6 <5.0.0",
+        "type": "range"
       },
-      "C:\\Users\\ITS-PCF-007\\Documents\\mariage"
+      "C:\\Users\\louis\\Documents\\mariage"
     ]
   ],
-  "_from": "@angular/platform-browser@4.3.0",
-  "_id": "@angular/platform-browser@4.3.0",
+  "_from": "@angular/platform-browser@>=4.4.6 <5.0.0",
+  "_id": "@angular/platform-browser@4.4.6",
   "_inCache": true,
   "_location": "/@angular/platform-browser",
   "_nodeVersion": "6.9.5",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/platform-browser-4.3.0.tgz_1500063689741_0.28597231209278107"
+    "tmp": "tmp/platform-browser-4.4.6.tgz_1508372365966_0.5037521154154092"
   },
   "_npmUser": {
     "name": "angular",
     "email": "angular-core+npm@google.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "@angular/platform-browser@4.3.0",
+    "raw": "@angular/platform-browser@^4.4.6",
     "scope": "@angular",
     "escapedName": "@angular%2fplatform-browser",
     "name": "@angular/platform-browser",
-    "rawSpec": "4.3.0",
-    "spec": "4.3.0",
-    "type": "version"
+    "rawSpec": "^4.4.6",
+    "spec": ">=4.4.6 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
-    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/platform-browser/-/platform-browser-4.3.0.tgz",
-  "_shasum": "02389489185185c3becf06359346100e5479c7e1",
+  "_resolved": "https://registry.npmjs.org/@angular/platform-browser/-/platform-browser-4.4.6.tgz",
+  "_shasum": "a9839c547e1b654fa1d24a89780c8ba6ab8dcce0",
   "_shrinkwrap": null,
-  "_spec": "@angular/platform-browser@4.3.0",
-  "_where": "C:\\Users\\ITS-PCF-007\\Documents\\mariage",
+  "_spec": "@angular/platform-browser@^4.4.6",
+  "_where": "C:\\Users\\louis\\Documents\\mariage",
   "author": {
     "name": "angular"
   },
@@ -59,8 +58,8 @@
   "devDependencies": {},
   "directories": {},
   "dist": {
-    "shasum": "02389489185185c3becf06359346100e5479c7e1",
-    "tarball": "https://registry.npmjs.org/@angular/platform-browser/-/platform-browser-4.3.0.tgz"
+    "shasum": "a9839c547e1b654fa1d24a89780c8ba6ab8dcce0",
+    "tarball": "https://registry.npmjs.org/@angular/platform-browser/-/platform-browser-4.4.6.tgz"
   },
   "es2015": "./@angular/platform-browser.js",
   "homepage": "https://github.com/angular/angular#readme",
@@ -76,8 +75,8 @@
   "name": "@angular/platform-browser",
   "optionalDependencies": {},
   "peerDependencies": {
-    "@angular/core": "4.3.0",
-    "@angular/common": "4.3.0"
+    "@angular/core": "4.4.6",
+    "@angular/common": "4.4.6"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
@@ -86,5 +85,5 @@
   },
   "scripts": {},
   "typings": "./platform-browser.d.ts",
-  "version": "4.3.0"
+  "version": "4.4.6"
 }
